+++ src/container.ts
@@ -1,5 +1,7 @@
 /**
  * Container options.
+ *
+ * @deprecated
  */
 export interface UseContainerOptions {
     /**
@@ -13,8 +15,14 @@ export interface UseContainerOptions {
     fallbackOnErrors?: boolean
 }
 
+/**
+ * @deprecated
+ */
 export type ContainedType<T> = { new (...args: any[]): T } | Function
 
+/**
+ * @deprecated
+ */
 export interface ContainerInterface {
     get<T>(someClass: ContainedType<T>): T
 }
@@ -22,6 +30,8 @@ export interface ContainerInterface {
 /**
  * Container to be used by this library for inversion control. If container was not implicitly set then by default
  * container simply creates a new instance of the given class.
+ *
+ * @deprecated
  */
 const defaultContainer: ContainerInterface = new (class
     implements ContainerInterface
@@ -47,6 +57,8 @@ let userContainerOptions: UseContainerOptions | undefined
 
 /**
  * Sets container to be used by this library.
+ *
+ * @deprecated
  */
 export function useContainer(
     iocContainer: ContainerInterface,
@@ -58,6 +70,8 @@ export function useContainer(
 
 /**
  * Gets the IOC container used by this library.
+ *
+ * @deprecated
  */
 export function getFromContainer<T>(someClass: ContainedType<T>): T {
     if (userContainer) {