+++ test/utils/test-utils.ts
@@ -397,7 +397,7 @@ export async function createTestingConnections(
 export function closeTestingConnections(connections: DataSource[]) {
     return Promise.all(
         connections.map((connection) =>
-            connection && connection.isConnected
+            connection && connection.isInitialized
                 ? connection.close()
                 : undefined,
         ),